Changed control index to float as I've seen this being returned from a page when exporting a page.

+++ Core/OfficeDevPnP.Core/Pages/ClientSidePageControls.cs
@@ -149,7 +149,7 @@ namespace OfficeDevPnP.Core.Pages
         /// </summary>
         /// <param name="controlIndex">The sequence of the control inside the section</param>
         /// <returns>Html representation of a control</returns>
-        public abstract string ToHtml(int controlIndex);
+        public abstract string ToHtml(float controlIndex);
 
         /// <summary>
         /// Removes the control from the page
@@ -330,7 +330,7 @@ namespace OfficeDevPnP.Core.Pages
         /// </summary>
         /// <param name="controlIndex">The sequence of the control inside the section</param>
         /// <returns>Html representation of this <see cref="ClientSideText"/> control</returns>
-        public override string ToHtml(int controlIndex)
+        public override string ToHtml(float controlIndex)
         {
             // Can this control be hosted in this section type?
             if (this.Section.Type == CanvasSectionTemplate.OneColumnFullWidth)
@@ -629,7 +629,7 @@ namespace OfficeDevPnP.Core.Pages
         /// </summary>
         /// <param name="controlIndex">The sequence of the control inside the section</param>
         /// <returns>HTML representation of the client side web part</returns>
-        public override string ToHtml(int controlIndex)
+        public override string ToHtml(float controlIndex)
         {
             // Can this control be hosted in this section type?
             if (this.Section.Type == CanvasSectionTemplate.OneColumnFullWidth)
